1. 방어막 구현 완료(방어막 카드 : shield, AmuletOfSteal) 2. 방어막 UI는 간단하게 옆에 숫자만 띄우자 3. 무적 방패 구현중 (invincible shield)

+++ Assets/Script/Controllers/Base/BaseController.cs
@@ -202,10 +202,10 @@ public abstract class BaseController : MonoBehaviourPun, IPunObservable
     }
 
     //퍼센트 계산
-    protected double PercentageCount(double percent, double value, int decimalplaces)
+    protected float PercentageCount(double percent, double value, int decimalplaces)
     {
 
-        return System.Math.Round(percent / 100 * value, decimalplaces);
+        return (float)System.Math.Round(percent / 100 * value, decimalplaces);
     }
